add context support
This allows for easy creation of context-aware logging filters without having to
pass the context as a "magic" value (kind of like `err` currently is XXX). The
main aim of this API breaking change is to allow ln log entries to automagically
have all of their values added to opentracing spans.

This will need changes to `ln-roll` and other packages that depend on `ln`.
This change breaks the public API (for good reasons!).